+++ crates/hc_demo_cli/src/demo.rs
@@ -150,7 +150,7 @@ async fn gen_dna_file(output: std::path::PathBuf) {
     .unwrap();
 
     let i_wasm = DnaWasmHashed::from_content(DnaWasm {
-        code: Arc::new(i_wasm.into_boxed_slice()),
+        code: i_wasm.into(),
     })
     .await;
     let i_zome = IntegrityZomeDef::from(ZomeDef::Wasm(WasmZome::new(i_wasm.hash.clone())));
@@ -163,7 +163,7 @@ async fn gen_dna_file(output: std::path::PathBuf) {
     .unwrap();
 
     let c_wasm = DnaWasmHashed::from_content(DnaWasm {
-        code: Arc::new(c_wasm.into_boxed_slice()),
+        code: c_wasm.into(),
     })
     .await;
     let c_zome = CoordinatorZomeDef::from(ZomeDef::Wasm(WasmZome::new(c_wasm.hash.clone())));